TO MUCH
Sooo, to start:  I added back the gravity button because we really don't need it until stage 4 & 5 and it was the only way I knew how to hide it until those stages. I added collision so that it now does account for sugar gravity. So when gravity is turned on, the sugar in a way follows the lines. It's not perfect to say but it works. After that, I did some general changes. Made sand falling and hitting cup stop playing during the next level screen. That way you can actually hear the win level screen audio. Made an actual game won screen, instead of having it just close out the window, and it allows us to use the gamewon audio.